Change IPntBin ID to long from int.
Also expand range, probably makes no difference.

+++ src/cfd_mesh/ISegChain.h
@@ -74,15 +74,15 @@ public:
 class IPntBin
 {
 public:
-    int m_ID;                           // x*1000 + y*1000 + z*1000
+    long m_ID;                           // x*1000 + y*1000 + z*1000
 
     deque< IPnt* > m_IPnts;
-    vector< int > m_AdjBins;            // Adjancent Bins
+    vector< long > m_AdjBins;            // Adjancent Bins
 
-    IPnt* Match( IPnt* ip, map< int, IPntBin > & binMap );
+    IPnt* Match( IPnt* ip, map< long, IPntBin > & binMap );
     void AddCompareIPnts( IPnt* ip, vector< IPnt* > & compareIPntVec );
 
-    static int ComputeID( vec3d & pos );
+    static long ComputeID( vec3d & pos );
 
 
 };
